printf

+++ include/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mhachem <mhachem@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/06 19:07:31 by mhachem           #+#    #+#             */
-/*   Updated: 2025/05/09 16:53:48 by mhachem          ###   ########.fr       */
+/*   Updated: 2025/05/09 17:27:20 by mhachem          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int		ft_putchar(char c);
 int		ft_putnbr(int nb);
 int		ft_putaddress(void *p);
 int		ft_handle_type(const char letter, va_list args);
-int		ft_puthexalower(unsigned int nb);
+int		ft_puthexalower(unsigned long nb);
 int		ft_puthexaupper(unsigned int nb);
 int		ft_putunsigned_dec(unsigned int nb);
