+++ srcs/ft_puthexalower.c
@@ -6,13 +6,13 @@
 /*   By: mhachem <mhachem@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/07 15:45:28 by mhachem           #+#    #+#             */
-/*   Updated: 2025/05/09 16:29:33 by mhachem          ###   ########.fr       */
+/*   Updated: 2025/05/09 17:27:14 by mhachem          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_puthexalower(unsigned int nb)
+int	ft_puthexalower(unsigned long nb)
 {
 	int		length;
 	char	*base;
